refactor(server/events): extract makeRepliesObserver to observers/replies.ts
Move the annotation-replies Y.Map observer body into a standalone factory.
Replies observer accesses both annotationsMap (parent lookup) and repliesMap
from doc inside the factory. Remove now-unused Y_MAP_ANNOTATIONS,
Y_MAP_ANNOTATION_REPLIES, Annotation, AnnotationReply imports from queue.ts.
